requirements.txt
playwright>=1.44.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
loguru>=0.7.2
httpx>=0.27.0
